Fix #58 revert `Aus (Bus)` to Uninomial (Author)
This is an unfortunate and hopefully temporary solution. We are
fixing botanical cases for these names but do not parse subgenus
data correctly. Still it is better to have 'fake' authours, than
'fake' subgenera. It seems the only way to solve this is to have
a dictionary of botanical authors and make a separate parser for
botanical cases.