Fix #64, fix #65 gRPC improvements
Add method that allows to send an array of names to gRPC. The array
is cleaned up and processed in several go-routines, then assembled
in the same order as intput array.

Move from serving JSON strings y gRPC to real objects. It probably
makes sense to use these gRPC objects internally when gnparser is
used as Go library.