remove config file (close #144)
gnparser does not have that many flags, and config file has a
limited value. If someone uses gnparser inside of a script
config creates additonal inconvenience. So I decided to remove
it for now. If there will be a flag explosion we might need
to get back to config file and think how to make it less
obtrusive.